wx_agent/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
datasets==2.18.0
dill==0.3.8
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
httpx-sse==0.3.1
huggingface-hub==0.21.3
ibm-generative-ai==2.2.0
idna==3.6
Janome==0.5.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.28
langchain-text-splitters==0.0.1
langsmith==0.1.14
logging==0.4.9.6
lxml==5.1.0
marshmallow==3.21.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
orjson==3.9.15
packaging==23.2
pandas==2.2.1
plac==1.4.3
portalocker==2.8.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
regex==2023.12.25
requests==2.31.0
responses==0.18.0
sacrebleu==2.4.0
safetensors==0.4.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
sumeval==0.2.2
tabulate==0.9.0
tenacity==8.2.3
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.38.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
unidic-lite==1.0.8
urllib3==2.2.1
uvicorn==0.27.1
xxhash==3.4.1
yarl==1.9.4
